results so far
Have run all analyses except for predictions 1 and 3 as need to calculate correlations and need to compute copied_prestigious correctly.
Also must check that our topic score and topic copies match up with those reported in the node tables. And think about controlling for topic in all models as well as ppt and group...